ai/requirements.txt
# WCAG-VLM Requirements
# Python 3.10+

# === CORE DEPENDENCIES ===
torch==2.1.1
torchvision==0.16.1


# Vision-Language Models
transformers==4.46.1
datasets==2.14.6  # For downloading WebSight
accelerate>=0.26.0
bitsandbytes==0.41.1  # For quantization

# Parameter-Efficient Fine-Tuning
peft==0.7.0  # LoRA, QLoRA support
scipy==1.11.4  # Required by peft

# === WEB SCRAPING ===
pyppeteer==1.0.2  # Async headless browser
selenium==4.15.2  # Alternative browser automation
aiohttp==3.9.1  # Async HTTP client
requests==2.31.0  # Synchronous HTTP client

# === DATA PROCESSING ===
pandas==2.1.3
numpy==1.26.2
pillow==10.1.0  # Image processing
opencv-python==4.8.1.78

# === EVALUATION & METRICS ===
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0  # Interactive visualizations

# === DATABASE & STORAGE ===
pyarrow==14.0.1  # Parquet file support
h5py==3.10.0  # HDF5 support

# === WEB API & INFERENCE ===
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# === MONITORING & LOGGING ===
wandb==0.16.1  # Weights & Biases integration
tensorboard==2.15.1
tqdm==4.66.1  # Progress bars

# === UTILITIES ===
python-dotenv==1.0.0  # .env file support
pyyaml==6.0.1  # YAML config parsing
click==8.1.7  # CLI framework
loguru==0.7.2  # Advanced logging

# === TESTING ===
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# === DOCUMENTATION ===
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# === HPC & DISTRIBUTED TRAINING ===
ray==2.10.0  # Distributed computing
ray[tune]==2.10.0  # Hyperparameter tuning
ray[serve]==2.10.0  # Model serving

# Optional but recommended:
# nvidia-nccl-cu12==2.18.1  # NCCL for multi-GPU
# horovod==0.28.1  # Alternative distributed training (if needed)
